+++ test_settings.py
@@ -113,6 +113,7 @@ INSTALLED_APPS = (
     "allauth.socialaccount.providers.foursquare",
     "allauth.socialaccount.providers.frontier",
     "allauth.socialaccount.providers.fxa",
+    "allauth.socialaccount.providers.garminconnect",
     "allauth.socialaccount.providers.gitea",
     "allauth.socialaccount.providers.github",
     "allauth.socialaccount.providers.gitlab",